+++ i18next.js
@@ -1,6 +1,6 @@
 (function (global, factory) {
 	typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
-	typeof define === 'function' && define.amd ? define('i18next', factory) :
+	typeof define === 'function' && define.amd ? define(factory) :
 	(global.i18next = factory());
 }(this, function () { 'use strict';